+++ Source/Additions/UIKit/UIImage+BFKit.m
@@ -4,7 +4,7 @@
 //
 //  The MIT License (MIT)
 //
-//  Copyright (c) 2014 - 2016 Fabrizio Brancati. All rights reserved.
+//  Copyright (c) 2014 - 2018 Fabrizio Brancati.
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -143,8 +143,14 @@
 }
 
 - (UIImage * _Nonnull)imageAtRect:(CGRect)rect {
-    CGImageRef imageRef = CGImageCreateWithImageInRect([self CGImage], rect);
-    UIImage *subImage = [UIImage imageWithCGImage:imageRef];
+    CGRect realRect;
+    if (rect.origin.x != 0 && rect.origin.y != 0) {
+        realRect = CGRectMake(0, 0, rect.size.width * self.scale, rect.size.height * self.scale);
+    } else {
+        realRect = CGRectMake(rect.origin.x * self.scale, rect.origin.y * self.scale, rect.size.width * self.scale, rect.size.height * self.scale);
+    }
+    CGImageRef imageRef = CGImageCreateWithImageInRect([self CGImage], realRect);
+    UIImage *subImage = [UIImage imageWithCGImage:imageRef scale:self.scale orientation:self.imageOrientation];
     CGImageRelease(imageRef);
     
     return subImage;