+++ Source/Additions/UIKit/UIScreen+BFKit.m
@@ -4,7 +4,7 @@
 //
 //  The MIT License (MIT)
 //
-//  Copyright (c) 2014 - 2016 Fabrizio Brancati. All rights reserved.
+//  Copyright (c) 2014 - 2018 Fabrizio Brancati.
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -46,11 +46,17 @@
 
 - (CGSize)fixedScreenSize {
     CGSize screenSize = self.bounds.size;
-    
-    if ((NSFoundationVersionNumber <= NSFoundationVersionNumber_iOS_7_1) && UIInterfaceOrientationIsLandscape([UIApplication sharedApplication].statusBarOrientation)) {
+
+    UIInterfaceOrientation orientation = UIInterfaceOrientationPortrait;
+    Class UIApplicationClass = NSClassFromString(@"UIApplication");
+    if(UIApplicationClass && [UIApplicationClass respondsToSelector:@selector(sharedApplication)]) {
+        orientation = [[[UIApplication class] performSelector:@selector(sharedApplication)] statusBarOrientation];
+    }
+
+    if ((NSFoundationVersionNumber <= NSFoundationVersionNumber_iOS_7_1) && UIInterfaceOrientationIsLandscape(orientation)) {
         return CGSizeMake(screenSize.height, screenSize.width);
     }
-    
+
     return screenSize;
 }